Endret navn på funkjsonen - kortNedOppgavebeskrivelse. Flyttet til felles utils fil. Rettet opp i skrivefeil på tester.
