+++ src/frontend/Komponenter/Oppgavebenk/OppgaveRad.tsx
@@ -9,10 +9,9 @@ import {
 } from '../../App/typer/behandlingstema';
 import { formaterIsoDato, formaterIsoDatoTid } from '../../App/utils/formatter';
 import { Popover, Table } from '@navikt/ds-react';
-import { utledetFolkeregisterIdent } from './utils';
+import { kortNedOppgavebeskrivelse, utledetFolkeregisterIdent } from './utils';
 import { OppgaveKnapp } from './OppgaveKnapp';
 import { TableDataCellSmall } from '../Personoversikt/HistorikkVedtaksperioder/vedtakshistorikkUtil';
-import { kortNedBeskrivelse } from './kortNedBeskrivelse';
 
 interface Props {
     oppgave: IOppgave;
@@ -76,7 +75,9 @@ const OppgaveRad: React.FC<Props> = ({ oppgave, mapper, settFeilmelding, hentOpp
                 <TableDataCellSmall>{typeBehandling}</TableDataCellSmall>
                 <TableDataCellSmall>{fristFerdigstillelseDato}</TableDataCellSmall>
                 <TableDataCellSmall>{prioritet}</TableDataCellSmall>
-                <TableDataCellSmall>{kortNedBeskrivelse(oppgave.beskrivelse)}</TableDataCellSmall>
+                <TableDataCellSmall>
+                    {kortNedOppgavebeskrivelse(oppgave.beskrivelse)}
+                </TableDataCellSmall>
                 <TableDataCellSmall>{utledetFolkeregisterIdent(oppgave)}</TableDataCellSmall>
                 <TableDataCellSmall>{oppgave.tildeltEnhetsnr}</TableDataCellSmall>
                 <TableDataCellSmall>{enhetsmappe}</TableDataCellSmall>
